Adopt to SAP WEB IDE
