+++ .che/project.json
@@ -1,1 +1,1 @@
-{"type":"sap.web","builders":{"configs":{}},"runners":{"configs":{}},"attributes":{"sap.watt.common.setting":["{\"projectType\":[\"sap.watt.uitools.ide.web\",\"sap.watt.uitools.ide.fiori\"]}"]},"mixinTypes":[]}
+{"type":"sap.web","builders":{"configs":{}},"runners":{"configs":{}},"attributes":{"sap.watt.common.setting":["{\"projectType\":[\"sap.watt.uitools.ide.web\",\"sap.watt.uitools.ide.fiori\"],\"hcpdeploy\":{\"account\":\"p1942861385trial\",\"name\":\"openui5genericapptesting\",\"lastVersionWeTriedToDeploy\":\"1.0.8\"},\"codeCheckingTriggers\":{\"notifyBeforePush\":true,\"notifyBeforePushLevel\":\"Error\",\"blockPush\":false,\"blockPushLevel\":\"Error\"},\"translation\":{\"translationDomain\":\"\",\"supportedLanguages\":\"\",\"defaultLanguage\":\"\",\"defaultI18NPropertyFile\":\"\",\"resourceModelName\":\"\"}}"]},"mixinTypes":[]}
