+++ .gitignore
@@ -2,3 +2,5 @@ node_modules/
 *.swp
 **/npm-debug.log
 che/
+/.che/
+/dist/
