+++ webapp/test/steps/iClickOnAggregationItem.js
