+++ webapp/test/utils/getComponentView.js
