+++ webapp/test/utils/positionTextToIndex.js
